refactor to seperate out friends service
